[aapt2] Use the aapt2 from `build-tools_r$(XABuildToolsVersion)-{os}.zip`
Context: https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1288832
Context: https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1287930

Commit c9e2d754 had to revert the version of `aapt2` we were shipping
due to issues with older PC's. Specifically a certain type of AMD
processor the AMD Phenom II. That commit simply reverted back to the last
known working version.

This made us think about if getting `aapt2` from maven was the best
idea. We are not 100% sure if the builds on maven even match those
being shipped in the google sdk. So the idea now is to use the one that
google ships with its sdk. However since we will control which version
of the build tools we want to ship we will not be relying on the one
from the users SDK folder. This will make sure we always ship a known
working version of `aapt2` with our product.

The `XABuildToolsVersion` value will be sourced from [Configuration.props](https://github.com/xamarin/xamarin-android/blob/main/Configuration.props#L120).
the `os` value will be `windows`, `linux` or `macosx`.
At the time of this commit the `XABuildToolsVersion` value was 30.0.2.