[ci] Uninstall apkdiff before installation
We've been running into issues trying to install an older version of
apkdiff on various machines after bumping the version in a PR.  Avoid
this by always attempting to uninstall apkdiff first.  The uninstall
exit code is ignored as uninstall attempts will error if the tool is not
already installed.