[One .NET] version .NET 6 packages with Android API level
Our current versions for .NET 6 packages are:

    11.0.200-preview.4.XXX

Where `XXX` is the git "commit distance".

This is the only place Android 11 remains, as we have been using the
Android API level for all other versions in .NET 6.

We should instead use a version number, such as:

    30.0.100-preview.4.XXX

This will likely make things less confusing, so this can match the
Android TFM as well:

    net6.0-android30