[Xamarin.Android.Build.Tasks] Add Support for AndroidJavaSource to Bindings
It would be nice to be able to introduce custom Java code and API's to a
binding project. This would allow developers to simplify an API and
still expose that to the C# Developer.

Using the existing `AndroidJavaSource` ItemGroup we can do exactly that.
This ItemGroup now supports the `Bind` metadata which will build and
bind the java code *prior* to the main C# compiled running. Historically
`AndroidJavaSource` was only available on Application Projects. However
this ItemGroup can now be used in any project type. The `Bind` metadata
is `true` by default and will cause the Java code to be compiled into
a .jar file before the C# compiler is run. It will then be bound using
the Java.Interop Java->C# binding system. The final C# apis will be
generated an included in the final C# compilation.

There are some restrictions on the Java code. You can only use Java types
or types which are defined in .jar or .aar files you reference for that
project via `AndroidLibrary` or `EmbeddedJar`.
It is advised that you DO NOT export types which include Java Generics
via this mechanism. Like normal bindings, Java Generics cause significant
problems. Try to stuck to normal types if you can.