Set $ANDROID_SDK_HOME appropriately
Previously, we were setting `$ANDROID_SDK_HOME` to the actual path of
our Android SDK, which causes `avdmanager` to print warnings such as:

    Warning: ANDROID_SDK_HOME is set to the root of your SDK: /Users/runner/Library/Android/sdk
    ANDROID_SDK_HOME is meant to be the path of the preference folder expected by the Android tools.

It appears that `avdmanager` expects `$ANDROID_SDK_HOME` to point to a
location where it will create a new `.android` folder with AVDs.

I created a new `$(AvdManagerHome)` property to be used by the
`<StartAndroidEmulator/>` and `<CreateAndroidEmulator/>` MSBuild
tasks. We now have full control of where the `.android` folder is
located, which will be `~/android-toolchain/.android` by default.