[runtime] Less reflection
Try to remove as much reflection (both managed and Java) from the native
startup sequence.

Instead of using full reflection for managed classes, this commit
implements class and field loading using token IDs.  The IDs are
gathered at the time when typemaps are generated and then stored in the
generated assembly code, so that the native runtime can simply pass them
to Mono in order to load the needed class and its members.

Additionally, I noticed that the default Java time zone is no longer
requested during startup, so I moved the class lookup to a lazy code
path, protected with a mutex.

The changes make startup (till the app is displayed) around 30ms faster
for plain Xamarin.Android running on Pixel 3 XL