[aapt2] Use the aapt2 from `build-tools_r$(XABuildToolsVersion)-{os}.zip` (#5888)
Context: https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1288832
Context: https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1287930

Commit c9e2d754 had to revert the version of `aapt2` we were shipping
due to issues with older PC's, specifically those with an
AMD Phenom II CPU.  Commit c9e2d754 simply reverted back to the last
known working version.

This made us think about if getting `aapt2` from maven was the best
idea.  We are not 100% sure if the builds on maven even match those
being shipped in Google's Android SDK.

The idea now is to use the `aapt2` that Google ships with the
Android SDK.  However, since we will control which version of the
Android SDK Build-tools we want to ship, we will *not* be relying on
the one from the users Android SDK folder.  This will ensure we
always ship a known working version of `aapt2` with our product.

The Android SDK Build-tools version we use will be the version
specified in the `$(XABuildToolsVersion)` MSBuild property within
`Configuration.props`, which is currently 30.0.2.

If `aapt2` continues to fail on certain customer machines, then we
can direct them to download the `aapt2` used in c9e2d754 and override
the `$(Aapt2ToolExe)` and `$(Aapt2ToolPath)` MSBuild properties,
which is described in `Documentation/release-notes/5888.md`.