[Xamarin.Android.Build.Tasks] Allow `package` to be defined with AndroidManifestPlaceholders.
Fixes #5951
Related: commit b423ff6.

Our current `AndroidManifestPlaceholders` implementation would allow users
to change any part of the manifest with the exception of the `package`
element on the `manifest` itself. What was actually happening was the
`package` was being changed but it was happening after  the
`GetAndroidPackageName` had already been called. So the internal MSBuild
property `$(_AndroidPackage)` did not reflect the desired value.

So lets update `GetAndroidPackageName` to use the `AndroidManifestPlaceholders`.
Note: This will only effect the `$(_AndroidPackage)` value when the
manifest is being used. If the user is using the `GenerateApplicationManifest`
option, the `$(_AndroidPackage)` will be controlled from the `$(ApplicationId)`
MSBuild property.