[One .NET] remove NuGet workarounds
Fixes: https://github.com/xamarin/xamarin-android/issues/5896

Since early in .NET 5 development, we had workarounds (hacks?) for
enabling the legacy Xamarin.Forms NuGet package to work correctly. We
used `$(PackageTargetFallback)` so that `monoandroid` assemblies would
be used in NuGet packages. We also had to fix up item groups manually
so `monoandroid` could be chosen over `netstandard`.

We should be able to remove these workarounds, since we now have a
dotnet/installer with this change:

https://github.com/NuGet/NuGet.Client/commit/10274281c37e9dd3626938757f7e947d649b821b