[tests] remove SingleThreadedAttribute
We occasionally hit test failures in the
`InstallAndRunTests.InterpreterEnabled` test. Running the test locally
by itself seems to work fine.

Then I noticed the class was decorated with the `[SingleThreaded]`
attribute and not `[NonParallelizable]`. `[SingleThreaded]`'s summary
says:

> Marks a test fixture as requiring all child tests to be run on the
> same thread as the OneTimeSetUp and OneTimeTearDown. A flag in the
> NUnit.Framework.Internal.TestExecutionContext is set forcing all
> child tests to be run sequentially on the current thread. Any
> NUnit.Framework.ParallelScope setting is ignored.

How is this test class not failing *more* often than it is? I found
that we set an assembly-level attribute in `DeploymentTest.cs`:

    [assembly: NonParallelizable]

But from the description of `[SingleThreaded]`, it seems like this
setting would be ignored.

To (hopefully) make things better:

* Move `[assembly: NonParallelizable]` to an easier to find
  `AssemblyInfo.cs` file.
* Remove all usage of `[NonParallelizable]` in other files in the
  `MSBuildDeviceIntegration` project.
* Remove all usage of `[SingleThreaded]`.
* Fix a couple places we had a `UsesDevices` category, as `UsesDevice`
  appears to be the category used more often.
* Add `[TestFixture]` in a few places that it was missing.