Update VSCode Support
Update the extensions list to use the correct id for the vscode csharp
extension.

Remove the `nxunitExplorer.nunit` setting as it was causing the unit
tests not to load. This is because the extension is expecting a CIL
image, but we were passing a batchfile. As a result none of the unit
tests would load in the IDE.

Add some documentation around using VSCode to run the unit tests and
debug the sample apps. This includes a link to a patched version of the
`testrun.exe` which the `nxunit.explorer` extension uses. There is a bug
in the extension where tests will parameters will NOT run or debug. The
link patched version of the files is included in the docs. A PR to the
upstream extension has also been sent.

Added the `dotnet-adapter` extension which lets us run the tests under a `dotnet`
context.