[Xamarin.Android.Build.Tasks] fix warning with $(UseInterpreter)
We currently emit a warning when `$(UseInterpreter)` is true:

    warning XA0119: Using fast deployment and AOT at the same time is not recommended. Use fast deployment for Debug configurations and AOT for Release configurations. [C:\src\xamarin-android\bin\TestDebug\temp\DotNetBuildXamarinFormsTrue\UnnamedProject.csproj]

It turns out that we were setting `$(AotAssemblies)` to true when the
interpreter was used.

Since `$(UseInterpreter)` on Android is meant for Debug & "Hot Reload"
scenarios, we definitely don't want this setting to enable AOT!

Luckily enabling AOT doesn't do anything yet in .NET 6, because it is
not implemented yet. I also suspect enabling the interpreter is not
used much (if at all) in "legacy" Xamarin.Android.

I updated a .NET 6 test so it sets `$(UseInterpreter)` and verifies
there are no build warnings.