Revert "[One .NET] remove NuGet workarounds" (#6026)
Context: https://github.com/dotnet/maui/pull/1322#issuecomment-861603825

This conditionally reverts commit ca51adbc.

There is a change in NuGet that is currently needed by the .NET 6
Android workload:

https://github.com/NuGet/NuGet.Client/commit/10274281c37e9dd3626938757f7e947d649b821b

This change is not present in VS 16.10 or VS 16.11:

https://github.com/dotnet/msbuild/blob/vs16.11/eng/Version.Details.xml

If we want Visual Studio 2019 on Windows to continue working for .NET
6 projects, then we can't actually remove our NuGet workarounds.

I put the workarounds back, but then are now conditionally imported:

    <Import Project="Microsoft.Android.Sdk.NuGet.targets" Condition=" '$(MSBuildRuntimeType)' == 'Full' and '$(MSBuildVersion)' &lt; '17.0' " />

So if you use `dotnet build` or Visual Studio 2022, these will not be
used. This way we can keep VS 2019 working for some period of time.