[Xamarin.Android.Build.Tasks] Add OpenTK/NUnitLite Error for .net 6 (#6031)
Rework of commit 064babf6 to raise an error under .NET 6 if the user tries to
use OpenTK or NUnitLite. We can do this by using the `AndroidError`
task to raise the error. Under legacy projects we use the `ContinueOnError`
Task property to control if the output is a warning or not. Under .NET 6
it will be an error, under "legacy" Xamarin.Android it will be a warning.

Update the existing unit test to handle the new errors/warnings and
let it run under .NET 6.