[templates] set .png files to copyOnly
Context: https://github.com/dotnet/templating/issues/3325
Context: https://github.com/dotnet/templating/wiki/Reference-for-template.json#content-manipulation

In the current bump, `dotnet new android` + `dotnet build` fails with:

    Resources\mipmap-xxxhdpi\ic_launcher_round.png error APT2000: failed reading from input: PNG chunk type 49444154 is too large: chunk length is 10836 but chunk starts at byte 45/8187.

The .NET templating system has a 'copyOnly' mode for files that do not
need any text replaced. This is a performance feature, but it also
happens to workaround this issue.

We should be doing this on `.png` files *anyway*, as we don't need
potential "text" to be replaced. I will probably need to make similar
changes in xamarin/xamarin-macios and dotnet/maui.