[tests] fix for DotNetNew tests
If you run the `DotNetNew` tests by themselves, they fail because a
`NuGet.config` file is not generated that contains:

     <add key="testsource1" value="C:\src\xamarin-android\bin\BuildDebug\nuget-unsigned" />

...and so it isn't able to locate any of the
`Microsoft.Android.Runtime.*` packs without this.

We have two `CreateDotNetBuilder()` methods in this test, and they now
generate the `NuGet.config` file in the same way.

I also fixed a place that `Directory.CreateDirectory()` was needed.