[Xamarin.Android.Build.Tasks] Add support for `android:useEmbeddedDex`
Fixes #5925
Context https://developer.android.com/topic/security/dex

Android 10 introduced a new attribute for the application element in the
AndroidManifest.xml, `android:useEmbeddedDex`. This allows the .dex files
to be run directly from the apk, rather than unpacked. This new attribute
does require that the .dex files be STORED in the apk rather than
COMPRESSED. Otherwise you get the following error

    Failure [INSTALL_FAILED_INVALID_APK: Some dex are not uncompressed and aligned correctly

So lets update BuildApk to use the `GetCompressionMethod` method to figure
out if we need to compress the .dex files. We can then update the `ReadAndroidManifest`
task to check for the new attribute and update the `AndroidStoreUncompressedFileExtensions`
ItemGroup accordingly.

Also added unit tests.