[One .NET] enable C# bindings for @(AndroidLibrary) (#6056)
Fixes: https://github.com/dotnet/maui/issues/1475

A .NET 6 application project was failing to include C# bindings for
`.aar` files, but `.jar` files were working.

The workaround was to include:

    <LibraryProjectZip Include="myjavalibrary.aar" />

We had a test that checked this behavior for `.jar` files, but not
`.aar` files.

After reproducing the behavior in a test, I found the
`_CategorizeAndroidLibraries` MSBuild target did not automatically
setup the `@(LibraryProjectZip)` item group in .NET 6 application
projects. `@(AndroidAarLibrary)` *was* being setup, which explained
why the Java code made it to `classes.dex`, but C# bindings were
missing.

Updated the `_CategorizeAndroidLibraries` MSBuild target for this case.