Revert "[One .NET] normalize case for $(TargetPlatformIdentifier) (#6025)"
This reverts commit 15f23105d69e6c4cc50636b3575e3b14474a7671.

Doing this, in order to solve an issue with the IDE when
`$(TargetPlatformIdentifier)` is proper cased as `Android`. The
XamarinVS extension is comparing an exact match for `android` to light
up the Android device/emulator selection dropdown.

We can leave this change in place on the `main` branch, but we should
probably revert this for .NET 6 Preview 6. This way, we have more IDE
versions working as they did before.

See: https://github.com/xamarin/XamarinVS/pull/12385