[NET6] Remove as much AppDomain use as possible
NET6+ no longer supports multiple AppDomains, there's always only a
single "root" AppDomain.  This allows us to drop bits of code which deal
with obtaining and setting the current AppDomain in various context.

This commit removes almost all uses of AppDomains from the native
Xamarin.Android runtime.  Unfortunately, it requires sometimes messy use
of the C++ preprocessor, but wherever it is possible and not overly
ugly, I try to separate the NET6 and "legacy" code completely, for
easier and less noisy future removal of the latter.