[One .NET] Remove as much AppDomain use as possible (#6071)
Context: b7224e5a16de5ad3516ddb3898f647249160f3b8

.NET 5+ no longer supports multiple AppDomains, there's always only a
single "root" AppDomain.  This allows us to drop bits of code which
deal with obtaining and setting the current AppDomain in various
contexts.

Removes almost all uses of AppDomains from the native Xamarin.Android
runtime for .NET 6+.  Unfortunately, it requires sometimes messy use
of the C++ preprocessor, but wherever it is possible and not overly
ugly, I try to separate the NET6 and "legacy" code completely, for
easier and less noisy future removal of the latter.