[One .NET] select defaults for App Bundles
Fixes #6059

Users will probably want to target more than one App Store. Google is now requiring the `aab` format for
Google Play Store uploads. Unfortunately this package format is not compatible with other stores.
Users can build their app twice producing an `aab` for one build and `apk` for another, but we should
try to make this a bit easier.

`bundle-tool` has the ability to create a universal apk from the `aab` file. So lets make use of that to
generate one along side the `aab`. We are introducing a new property `AndroidPackageFormats`.
Under .net 6 this will have a value of `aab;apk` by default for Release builds, for Legacy it will be
empty.

If `AndroidPackageFormats` is specified and `AndroidPackageFormat` is empty we will use the
values in `AndroidPackageFormats` to populate `AndroidPackageFormat`. If none of these values
are provided the `AndroidPackageFormat` will still default to `apk` as it has always done.

The following setting will produce both an `aab` and an `apk`.
```
<PropertyGroup>
  <AndroidPackageFormats>aab;apk</AndroidPackageFormats>
</PropertyGroup>
```

This will produce just an `aab`

```
<PropertyGroup>
  <AndroidPackageFormats>aab</AndroidPackageFormats>
</PropertyGroup>
```

and this will produce just an `apk`

```
<PropertyGroup>
  <AndroidPackageFormats>apk</AndroidPackageFormats>
</PropertyGroup>
```

For .net 6 users this will be enabled by default for Release builds. For legacy users they will need
to define the property manually.