[One .NET] Update .NET 6 defaults for bindings projects (#6066)
Fixes: https://github.com/xamarin/xamarin-android/issues/4657

With .NET 6 we want to turn on the following binding project
enhancements added with C# 8 by default:

  - Interface constants: xamarin/java.interop@a3b74568
  - Interface static methods: xamarin/java.interop@855b7e92
  - Interface default methods: xamarin/java.interop@29f97075
  - Interface nested types: xamarin/java.interop@28b1fc97
  - Nullable Reference Types: xamarin/java.interop@01d06845

However, each of these features previously had a workaround that we
performed so the code could be used by .NET.  This means that users
who are updating their existing bindings may find API breaks as we
now use the new method instead of the workaround.  Thus each of these
needs a way for the user to opt out of the new behavior to preserve
API compatibility.

~~ New MSBuild Properties ~~

  - `$(AndroidBoundInterfacesContainConstants)`: Controls whether
    constants on interfaces will be supported, or the old workaround
    of creating a `IMyInterfaceConsts` class will be used.

  - `$(AndroidBoundInterfacesContainTypes)`: Controls whether types
    nested in interfaces will be supported, or the old workaround of
    creating a non-nested type like `IMyInterfaceMyNestedClass`.
  
  - `$(AndroidBoundInterfacesContainStaticAndDefaultInterfaceMethods)`:
    Controls whether default and static members on interfaces will be
    supported, or the old workaround of creating a sibling class
    containing static members like `abstract class MyInterface`.

~~ .NET 6 ~~

For .NET 6, the above properties all default to `true` if not
explicitly set by the user.

~~ Legacy ~~

For legacy, the above properties all default to `false`.
The "legacy" way of enabling these features was a single preview flag
`$(_EnableInterfaceMembers)`.  We will continue to support this flag
for legacy only, and it will enable all the above flags, unless a user
has explicitly set one of the flags.

~~ AndroidCodegenTarget ~~

.NET 6 binding projects will only support `XAJavaInterop1`, so other
values are ignored.