Remove $(SelfContained) = true by default
Builds were failing with:

    error NETSDK1031: It is not supported to build or publish a self-contained application without specifying a RuntimeIdentifier. You must either specify a RuntimeIdentifier or set SelfContained to false.

This seemed to be caused by reordering such as:

https://github.com/dotnet/sdk/pull/18639

In the past, we weren't actually overwriting `$(SelfContained)` *at
all*, and with these changes we now are. I set this value in the days
of .NET 5.

We should just remove where we set `$(SelfContained)` and let the
dotnet/sdk manage this value.