[CI] Add nuget to msi conversion and VS insert stage (#6030)
Context: https://github.com/xamarin/yaml-templates/pull/117

A new CI job and stage have been added to facilitate the
Visual Studio installer insertion process.

The first job named "Convert NuGet to MSI" will convert signed
packages to `.msi` installers, and generate Visual Studio manifests
for the `.msi` installers.  The `.msi` files are signed and converted
back to `.nupkg` files for the Windows command line install scenario.
These packages are uploaded as pipeline artifacts named
`vsdrop-signed`, and  `vs-msi-nugets`.

The new `vs-workload.props` file contains version information and other
metadata used by the convert job to generate a Visual Studio manifest.
The `@(ShortNames)` item group is used to avoid path length issues
that can arise due to the way VS caches some components when building.
The `.msi` conversion tooling uses this to replace strings in the VS
component file names and IDs generated for our workload packs.
In some cases the semver label generated for our packs can be long,
so we replace the full semver version with a more predictable four
part version (30.0.100.x).

The new "VS Insertion" stage starts with a [manual validation task][0].
This task will pause execution and wait for someone to click a "Resume"
or "Reject" button that will appear on the pipeline UI.  This task is
configured to be rejected after waiting for two days, but it can be
manually re-ran any time after being rejected to trigger a VS insertion
for an older build.

To access the Reume or Reject button, go to the Pipelines page for
the commit to be approved, e.g.:

 1. https://github.com/xamarin/xamarin-android/commits/main

 2. Click the ✅ or ❌ for the commit of interest.  This displays an
    "All checks have passed" or "Some checks were not successful"
    overlay.

 3. Click the **Details** link for **Xamarin.Android** within the
    overlay shown by (2).

 4. At the bottom of the "runs" page, click the
    **View more details on Azure Pipelines** link.

When the manual validation task is approved, a VS Drop will be
created containing all .NET 6 `.msi` files.  This Drop URL can then
be used to update our component versions in Visual Studio.  This last
piece is currently manual as we will initially be introducing new
components, however we should be able to automate VS PR creation in
the future.

[0]: https://docs.microsoft.com/en-us/azure/devops/pipelines/tasks/utility/manual-validation?view=azure-devops&tabs=yaml