[xaprepare] don't install microsoft-net-runtime-android workload (#6114)
Context: https://github.com/xamarin/xamarin-android/pull/6112#pullrequestreview-712229556

The .NET Preview 7 bump is failing with:

	Running: /Users/builder/Library/Android/dotnet/dotnet "workload" "install" "microsoft-net-runtime-android" "--skip-manifest-update" "--verbosity" "diag"
	stderr | Workload with id microsoft-net-runtime-android is not recognized.

However, it was a bug in the `dotnet workload install` command that
allowed us to even install this workload in the first place!
The concept concept behind "abstract" workloads is they are not user
visible in any way; they're meant to be an implementation detail.

In commit 0150bbbc, I setup `xaprepare` to install the
`microsoft-net-runtime-android` workload, but we don't actually even
need to do this.  We can simply update the `sdk-manifest` for this
pack, so that the version matches `$(DotNetRuntimePacksVersion)`.

The `microsoft-net-runtime-android` workload will get installed later
during the `ExtractWorkloadPacks` target, as part of installing the
`android` or `android-aot` workloads:

https://github.com/xamarin/xamarin-android/blob/c1a2ee70214e86757541b5759c9ed54941bd4680/build-tools/create-packs/Directory.Build.targets#L119-L122

Since `android` and `android-aot` extend
`microsoft-net-runtime-android`, this step will install it.