[Xamarin.Android.Build.Tasks] NullReferenceException in Xamarin.Android.Common.targets Fixes #6124
Commit 2a04748d removed the `[Required]` attribute from the `KeyPass`
and `StorePass` properties. This was because they were now optional.
However `keytool` does REQUIRE these parameters when we use the `-ks`
or `KeyStore` argument.

If a user does not pass these arguments they get the following error

```
    error XAAAS7001: System.NullReferenceException: Object reference not set to an instance of an object.
    error XAAAS7001: at Xamarin.Android.Tasks.AndroidApkSigner.AddStorePass(CommandLineBuilder cmd, String cmdLineSwitch, String value)
    error XAAAS7001: at Xamarin.Android.Tasks.AndroidApkSigner.GenerateCommandLineCommands()
```

This is because the `KeyPass` or `StorePass` are `null` and the `AddStorePass`
function is not handling that.

So what we should do is when the `KeyStore` argument is being used we should
error out if `KeyPass` or `StorePass` are empty or `null`.