[macOS] Do not pin mingw version
Homebrew allows pinning of a package to a specific version of it, but
the solution is far from perfect for us.  The problem is that if you
updated your mingw version to any version other than we require,
`xaprepare` will detect it and will try to install version of the
package that we need, but if it's older than the current version, we'll
get the latest version installed anyway since there's no direct way of
specifying a package/formula version.

Instead of trying o work around the issue, let's always install the
latest version of mingw (currently 9.0.0).