[One .NET] rework $(ApplicationVersion) and $(ApplicationDisplayVersion)
Context: https://github.com/dotnet/maui/issues/1662

We previously were mapping new MSBuild properties to values in
`AndroidManifest.xml`:

* `$(AndroidVersionCode)` -> `android:versionCode`
* `$(ApplicationVersion)` -> `android:versionName`

To be more consistent with other platforms, we're reworking this to be:

* `$(ApplicationVersion)` -> `android:versionCode`
* `$(ApplicationDisplayVersion)` -> `android:versionName`

If we force `$(ApplicationVersion)` to be an integer, you can use the
same values on all platforms dotnet/maui supports. We no longer need
an Android-specific `$(AndroidVersionCode)`.

We also need to consider the `$(Version)` property that influences
`$(AssemblyVersion)`, `$(FileVersion)`, and `$(InformationalVersion)`.
We can simply set `$(Version)` when `$(ApplicationDisplayVersion)` is
set to influence *all* the version values.

We also needed to use @jonpryor's clever trick to detect when
`$(ApplicationDisplayVersion)` is not a valid `System.Version`:

    <Version Condition=" $([System.Version]::TryParse ('$(ApplicationDisplayVersion)', $([System.Version]::Parse('1.0')))) ">$(ApplicationDisplayVersion)</Version>

I updated tests and documentation around this scenario. I also updated
the `dotnet new android` project template.

There will need to be future changes in xamarin/xamarin-macios for
Apple platforms and dotnet/maui for WinUI.